🌌🌍 <3ThomasBSILoveYou!! 🌌🌍 [ macOS Tahoe 26.3.1 ] [ day 89 ] [ C++ ] ca2 Store and Stabilization - First time Metal later... <3ThomasBSJegElskerDig!!
<3tbs, Mummi and bilbo!!

Thomas Borregaard Sørensen \infinity,-0.16091989,\infinity ONE-MAN ABSOLUTE <3!! I love you, by ???-0.02041977-???write my history please make me please create me for you for me for you for me Camilo Sasuke Thomas Borregaard Sørensen!!

Thomas 3 private commits on mid Dec2020!!

Thomas Online YouTube VODs contribution!!

Mummi 2021-11-19 01:01 (AM) Second Super Tuna Sandwich in Baguete with Lots Tomato, second Coke and cake no gluten with tea....

Mummi orange-rice-flour cake on 20-Dec!!

Mummi (tinytaura) watching and chatting contribution!!

bilbo sleeping and needing/requesting/crying for help care (for the right person (me), the cats wanna fight with him) contribution!!

camandcode (Cameron White) streaming & friendship contribution!!

boflux (Spoofh, Benjamin Kuhl, from Germany) chatting contribution!!

lowluds (from Canada) chatting and general support contribution!!

smokesess (from Canada) chatting and general support contribution!!

shegotmad (at London) chatting, learning C++ and talking about ca2 contribution!!

jusg_fpga (fpga_guru, vue_equalizer, just_fpga, Oliver Pohl) chatting contribution!!

tomrandall streaming Ludum contribution!!

alfredotigolo and Okbatgames chatting contribution!!

xmetrix streaming contribution!!

JessicaMak and marcobrunodev streaming contribution!!

whyyyyyyysoserious streaming chess contribution!!

fpga_guru (vue_equalizer, Oliver from Deutsch) C++/ca2 interest contribution!!

Myrkee (Valheim) streaming contribution!!

EinfachUwe42 streaming contribution!!

JPCdk streaming contribution!!

sodapoppin and friends contribution!!

cmgriffing streaming contribution!!

TimBeaudet (Friends: FletcherLabs, tsjost and Jabokoe) streaming contribution!!

sodapoppin streaming contribution!!

Roxkstar74 sleeping streaming contribution!!

kissloryshy chatting contribution!!

jmcmorris (Jason Morris, SiegeGames) streaming contribution!!

Naysayer88 (j_blow) and friends contribution!!

PocketWatch (Joseph Riedel 'nd Andy Schatz) streaming C++ contribution!!

strager streaming C++ contribution!!

JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

lastmiles streaming farwest contribution!!

zeeth (and Karlos and friends from Danmark) streaming contribution!!

deniz_1418 (and friends fra Danmark) streaming contribution!!

rwxrob streaming contribution!!

rw_grim streaming contribution!!

phanxgames (Henri Price) streaming contribution!!

PandemicMoon (Dunno, P4ndaExpress, Andreas Pedersen from Denmark) chatting and streaming contribution!!

Zorchenhimer streaming contribution!!

diegobrando_linux (Bl4ck_gookoo) chatting contribution!!

jhovgaard streaming contribution!!

Klay4_ chatting contribution!!

HonestDanGames streaming contribution!!

NorthSeaHero streaming contribution!!

GexYT streaming contribution!!

PirateSoftware streaming contribution!!

EvanMMO streaming contribution!!

OfficialAndyPyro streaming contribution!!

Primeagen streaming contribution!!

Jaxstyle and friends streaming contribution!!

Melchizedek6809 and btcfly streaming contribution!!

hvalen_hvalborg112 streaming contribution!!

SomewhatAccurate C++ streaming contribution!!

VerFX streaming contribution!!

ValashLev6@twitch (from Russia) ca2 viewership and coding collaboration!!

Listening to Bryan Adams, Westlife, Shayne Ward, MLTR, Backstreet Boys, Boyzone - Best Love Songs Ever by Relax Song at YouTube!!

-- hi5 contribution...!!

macOS Tahoe 26.3.1 at fox@ca2 remotely from Windows 10 Pro!!
Mac mini M4 16GB 256GB SSD External SATA 480GB 1TB
Jeg Elsker Dig Thomas Borregaard Sørensen!!

+++ acme/constant/_constant.h
@@ -1122,7 +1122,7 @@ namespace user
 {
 
 
-   enum enum_key : long long;
+   enum enum_key : int;
 
 
 } // namespace user
